refactor server socket handling
